+++ PSModuleDevelopment/changelog.md
@@ -1,6 +1,7 @@
 ﻿# Changelog
 ##
  - New: Format-PSMDParameter - updates legacy parameter notation
+ - New: Measure-PSMDLinesOfCode - Measures the lines of code in a scriptfile.
  - New: Search-PSMDPropertyValue - search objects for values in properties
  - Upd: Template PSFTest - adding WMI commands to list of forbidden commands
  - Upd: Template PSFModule - adding changelog